Handle nonexistent projects in CleanProjectWorker (#1822)
We have an acceptance tests that ensures cleaning works. This tests
started failing since we made the action asynchronous since on the
client we send a "delete project" right after the completion of the
clean call, leading to a race condition where by the time the job is
picked up by Oban, the project doesn't exist.

This race condition is very unlikely to reproduce in production, but I
added some logic to handle it gracefully anyways.

+++ lib/tuist/projects/workers/clean_project_worker.ex
@@ -10,12 +10,19 @@ defmodule Tuist.Projects.Workers.CleanProjectWorker do
   @impl Oban.Worker
   def perform(%Oban.Job{args: %{"project_id" => project_id}} = _job) do
     project = Tuist.Projects.get_project_by_id(project_id)
-    project_slug = "#{project.account.name}/#{project.name}"
 
-    Task.await_many([
-      Task.async(fn -> Storage.delete_all_objects("#{project_slug}/builds") end),
-      Task.async(fn -> Storage.delete_all_objects("#{project_slug}/tests") end),
-      Task.async(fn -> CacheActionItems.delete_all_action_items(%{project: project}) end)
-    ])
+    if is_nil(project) do
+      :ok
+    else
+      project_slug = "#{project.account.name}/#{project.name}"
+
+      Task.await_many([
+        Task.async(fn -> Storage.delete_all_objects("#{project_slug}/builds") end),
+        Task.async(fn -> Storage.delete_all_objects("#{project_slug}/tests") end),
+        Task.async(fn -> CacheActionItems.delete_all_action_items(%{project: project}) end)
+      ])
+
+      :ok
+    end
   end
 end
